Fixed refresh token
