+++ api/config/packages/lexik_jwt_authentication.yaml
@@ -2,7 +2,7 @@ lexik_jwt_authentication:
     secret_key: '%env(resolve:JWT_SECRET_KEY)%'
     public_key: '%env(resolve:JWT_PUBLIC_KEY)%'
     pass_phrase: '%env(JWT_PASSPHRASE)%'
-    token_ttl: 3600
+    token_ttl: 3600 # 1 hour
     api_platform:
         enabled: true
     token_extractors:
