+++ front/app/routes/_main/map/$placeId_.$heistId.delete/route.tsx
@@ -5,8 +5,9 @@ import { deleteHeist, heistIsMadeBy } from '~/lib/api/heist';
 import { i18next } from '~/lib/i18n/index.server';
 import { commitSession, getSession } from '~/lib/session.server';
 import { getMessageForErrorStatusCodes, hasErrorStatusCodes } from '~/lib/utils/api';
+import { ROLES } from '~/lib/utils/roles';
 import { FLASH_MESSAGE_KEY } from '~/root';
-import { ROLES, denyAccessUnlessGranted, hasRoles } from '~utils/security.server';
+import { denyAccessUnlessGranted, hasRoles } from '~utils/security.server';
 
 import type { ActionFunctionArgs } from '@remix-run/node';
 import type { FlashMessage } from '~/root';
